feat(threads): switch to wasi-threads
The new way of spinning threads in wasm is through WASI-Threads,
not with WAMR libpthread implementation. We switch it here.

The -Wl,--max-memory=1048576 parameter is necessary, because lld
will set it as default to 0 and not as unlimited, as stated in its
documentation.

We might need to adjust the maximum memory to be higher, but 1MB
seems enough for now.

Each thread needs 2 pages of memory (so 2x64KB). With 1MB we can
have 8 threads.

Signed-off-by: Marco Casaroli <marco.casaroli@gmail.com>
Signed-off-by: Dan Kouba <dan@atym.io>